Support a better name for the starter header
